convert provided mapping into JSON
