+++ src/link-arrows/left.json
@@ -1,0 +1,34 @@
+{
+	"0": "<:LinkMarker_LeftMiddle_0:1102794510444011581>",
+	"1": "<:LinkMarker_LeftMiddle_1:1102794512960589895>",
+	"2": "<:LinkMarker_LeftMiddle_12:1102794524545273879>",
+	"12": "<:LinkMarker_LeftMiddle_2:1102794515661725716>",
+	"4": "<:LinkMarker_LeftMiddle_4:1102794519281405963>",
+	"14": "<:LinkMarker_LeftMiddle_41:1102794527913295962>",
+	"24": "<:LinkMarker_LeftMiddle_42:1102794530643771505>",
+	"124": "<:LinkMarker_LeftMiddle_412:1102794550000504861>",
+	"7": "<:LinkMarker_LeftMiddle_7:1102794520774574110>",
+	"17": "<:LinkMarker_LeftMiddle_71:1102794532355055687>",
+	"27": "<:LinkMarker_LeftMiddle_72:1102794535404326922>",
+	"127": "<:LinkMarker_LeftMiddle_712:1102796157912096808>",
+	"47": "<:LinkMarker_LeftMiddle_74:1102794536851361792>",
+	"147": "<:LinkMarker_LeftMiddle_741:1102794555071402005>",
+	"247": "<:LinkMarker_LeftMiddle_742:1102794556669448302>",
+	"1247": "<:LinkMarker_LeftMiddle_7412:1102794572989481033>",
+	"8": "<:LinkMarker_LeftMiddle_8:1102794523463131236>",
+	"18": "<:LinkMarker_LeftMiddle_81:1102794541859356763>",
+	"28": "<:LinkMarker_LeftMiddle_82:1102794545877483610>",
+	"128": "<:LinkMarker_LeftMiddle_812:1102794567067127848>",
+	"48": "<:LinkMarker_LeftMiddle_84:1102794548532498503>",
+	"148": "<:LinkMarker_LeftMiddle_841:1102796162228035634>",
+	"248": "<:LinkMarker_LeftMiddle_842:1102796164748820550>",
+	"1248": "<:LinkMarker_LeftMiddle_8412:1102796165872898110>",
+	"78": "<:LinkMarker_LeftMiddle_78:1102794539799945266>",
+	"178": "<:LinkMarker_LeftMiddle_781:1102794559844532324>",
+	"278": "<:LinkMarker_LeftMiddle_782:1102796160625819678>",
+	"1278": "<:LinkMarker_LeftMiddle_7812:1102794574356811847>",
+	"478": "<:LinkMarker_LeftMiddle_784:1102794563749429279>",
+	"1478": "<:LinkMarker_LeftMiddle_7841:1102794577762603018>",
+	"2478": "<:LinkMarker_LeftMiddle_7842:1102794580648276069>",
+	"12478": "<:LinkMarker_LeftMiddle_78412:1102794587350782022>"
+}
