+++ src/link-arrows/right.json
@@ -1,0 +1,10 @@
+{
+	"0": "<:LinkMarker_Right_0:1102794415149432992>",
+	"3": "<:LinkMarker_Right_3:1102794418290958336>",
+	"6": "<:LinkMarker_Right_6:1102794419855429672>",
+	"36": "<:LinkMarker_Right_63:1102794424213323776>",
+	"9": "<:LinkMarker_Right_9:1102794422506238103>",
+	"39": "<:LinkMarker_Right_93:1102794426994147338>",
+	"69": "<:LinkMarker_Right_96:1102794428445368360>",
+	"369": "<:LinkMarker_Right_963:1102794429510733835>"
+}
